# docs/openapi.json
{
  "components": {
    "schemas": {
      "AdmissionCancelRequest": {
        "properties": {
          "reservation_token": {
            "minLength": 1,
            "title": "Reservation Token",
            "type": "string"
          },
          "run_id": {
            "exclusiveMinimum": 0.0,
            "title": "Run Id",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "reservation_token"
        ],
        "title": "AdmissionCancelRequest",
        "type": "object"
      },
      "AdmissionCancelResponse": {
        "properties": {
          "elder_id": {
            "title": "Elder Id",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status",
          "elder_id",
          "message"
        ],
        "title": "AdmissionCancelResponse",
        "type": "object"
      },
      "AdmissionConfirmRequest": {
        "properties": {
          "reservation_token": {
            "minLength": 1,
            "title": "Reservation Token",
            "type": "string"
          },
          "run_id": {
            "exclusiveMinimum": 0.0,
            "title": "Run Id",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "reservation_token"
        ],
        "title": "AdmissionConfirmRequest",
        "type": "object"
      },
      "AdmissionConfirmResponse": {
        "properties": {
          "bed_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bed Id"
          },
          "elder_id": {
            "title": "Elder Id",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status",
          "elder_id",
          "message"
        ],
        "title": "AdmissionConfirmResponse",
        "type": "object"
      },
      "AdmissionPreviewRequest": {
        "properties": {
          "elder_id": {
            "exclusiveMinimum": 0.0,
            "title": "Elder Id",
            "type": "integer"
          },
          "preferred_bed_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Bed Id"
          }
        },
        "required": [
          "elder_id"
        ],
        "title": "AdmissionPreviewRequest",
        "type": "object"
      },
      "AdmissionPreviewResponse": {
        "properties": {
          "bed_id": {
            "title": "Bed Id",
            "type": "integer"
          },
          "bed_no": {
            "title": "Bed No",
            "type": "string"
          },
          "elder_name": {
            "title": "Elder Name",
            "type": "string"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "preview_summary": {
            "title": "Preview Summary",
            "type": "string"
          },
          "reservation_token": {
            "title": "Reservation Token",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status",
          "elder_name",
          "bed_id",
          "bed_no",
          "preview_summary",
          "reservation_token",
          "expires_at"
        ],
        "title": "AdmissionPreviewResponse",
        "type": "object"
      },
      "AgentChatRequest": {
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "AgentChatRequest",
        "type": "object"
      },
      "AgentChatResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          }
        },
        "required": [
          "answer"
        ],
        "title": "AgentChatResponse",
        "type": "object"
      },
      "AlertAnalysisRequest": {
        "properties": {
          "alert_id": {
            "exclusiveMinimum": 0.0,
            "title": "Alert Id",
            "type": "integer"
          }
        },
        "required": [
          "alert_id"
        ],
        "title": "AlertAnalysisRequest",
        "type": "object"
      },
      "AlertAnalysisResponse": {
        "properties": {
          "alert_id": {
            "title": "Alert Id",
            "type": "integer"
          },
          "analysis": {
            "title": "Analysis",
            "type": "string"
          }
        },
        "required": [
          "alert_id",
          "analysis"
        ],
        "title": "AlertAnalysisResponse",
        "type": "object"
      },
...
      "CarePlanRequest": {
        "properties": {
          "care_goal": {
            "default": "",
            "maxLength": 1000,
            "title": "Care Goal",
            "type": "string"
          },
          "elder_name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Elder Name",
            "type": "string"
          }
        },
        "required": [
          "elder_name"
        ],
        "title": "CarePlanRequest",
        "type": "object"
      },
      "CarePlanResponse": {
        "properties": {
          "care_goal": {
            "title": "Care Goal",
            "type": "string"
          },
          "elder_name": {
            "title": "Elder Name",
            "type": "string"
          },
          "plan": {
            "title": "Plan",
            "type": "string"
          }
        },
        "required": [
          "elder_name",
          "care_goal",
          "plan"
        ],
        "title": "CarePlanResponse",
        "type": "object"
      },
...
      "CheckInRecommendationRequest": {
        "properties": {
          "elder_name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Elder Name",
            "type": "string"
          }
        },
        "required": [
          "elder_name"
        ],
        "title": "CheckInRecommendationRequest",
        "type": "object"
      },
      "CheckInRecommendationResponse": {
        "properties": {
          "elder_name": {
            "title": "Elder Name",
            "type": "string"
          },
          "suggestion": {
            "title": "Suggestion",
            "type": "string"
          }
        },
        "required": [
          "elder_name",
          "suggestion"
        ],
        "title": "CheckInRecommendationResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ZZYL Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agent/admission/cancel": {
      "post": {
        "operationId": "admission_cancel_api_agent_admission_cancel_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdmissionCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionCancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admission Cancel",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/agent/admission/confirm": {
      "post": {
        "operationId": "admission_confirm_api_agent_admission_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdmissionConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionConfirmResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admission Confirm",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/agent/admission/preview": {
      "post": {
        "operationId": "admission_preview_api_agent_admission_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdmissionPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admission Preview",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/agent/alert-analysis": {
      "post": {
        "operationId": "analyze_alert_api_agent_alert_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Alert",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/agent/care-plan": {
      "post": {
        "operationId": "generate_care_plan_api_agent_care_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarePlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Care Plan",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/agent/chat": {
      "post": {
        "operationId": "chat_api_agent_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/agent/chat/stream": {
      "post": {
        "operationId": "chat_stream_api_agent_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Server-sent event stream of chat tokens."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Stream",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/agent/checkin/recommendation": {
      "post": {
        "operationId": "recommend_checkin_api_agent_checkin_recommendation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInRecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInRecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recommend Checkin",
        "tags": [
          "agent"
        ]
      }
    },
...
  }
}